операции с деньгами - BigDecimal

+++ marketplace-order-service/src/main/java/org/yrti/order/dto/PricingResponse.java
@@ -4,12 +4,14 @@ import jakarta.validation.constraints.Min;
 import jakarta.validation.constraints.NotNull;
 import lombok.Data;
 
+import java.math.BigDecimal;
+
 @Data
 public class PricingResponse {
     @NotNull(message = "Product ID is required")
     private Long productId;
     @Min(value = 1, message = "Price must be at least 1")
-    private Double originalPrice;
+    private BigDecimal originalPrice;
     @Min(value = 1, message = "Price must be at least 1")
-    private Double discountedPrice;
+    private BigDecimal discountedPrice;
 }
